refactor: rename form component file to simplify import paths
